+++ src/export/mod.rs
@@ -5,11 +5,14 @@ mod asciidoc;
 mod csv;
 mod json;
 mod markdown;
+mod markup;
+mod orgmode;
 
 use self::asciidoc::AsciidocExporter;
 use self::csv::CsvExporter;
 use self::json::JsonExporter;
 use self::markdown::MarkdownExporter;
+use self::orgmode::OrgmodeExporter;
 
 use crate::benchmark::benchmark_result::BenchmarkResult;
 use crate::util::units::Unit;
@@ -31,6 +34,9 @@ pub enum ExportType {
 
     /// Markdown table
     Markdown,
+
+    /// Emacs org-mode tables
+    Orgmode,
 }
 
 /// Interface for different exporters.
@@ -66,6 +72,7 @@ impl ExportManager {
             add_exporter("export-json", ExportType::Json)?;
             add_exporter("export-csv", ExportType::Csv)?;
             add_exporter("export-markdown", ExportType::Markdown)?;
+            add_exporter("export-orgmode", ExportType::Orgmode)?;
         }
         Ok(export_manager)
     }
@@ -80,6 +87,7 @@ impl ExportManager {
             ExportType::Csv => Box::new(CsvExporter::default()),
             ExportType::Json => Box::new(JsonExporter::default()),
             ExportType::Markdown => Box::new(MarkdownExporter::default()),
+            ExportType::Orgmode => Box::new(OrgmodeExporter::default()),
         };
         self.exporters.push(ExporterWithFilename {
             exporter,